Auto generating simple arithmetic calculus questions for secondary schools using Python and Xlsx Writer.
Reason for this project: Taught arithmetic in a secondary school and was
given the task to set daily multiplication drill questions for junior
secondary school students over the course of two weeks. Wrote code to
automate the process because, no time for stress. I plan to build a GUI
for it using tkinter and expand functionality in the future. You will
need to install the xlsxwriter module to get this to work, since a bulk
of the program depends on it.